refactor(sdk): Move search index related code into its own module.
