+++ crates/matrix-sdk/src/event_cache/mod.rs
@@ -75,8 +75,6 @@ use crate::{
 mod deduplicator;
 mod pagination;
 mod room;
-#[cfg(feature = "experimental-search")]
-mod search;
 
 pub use pagination::{RoomPagination, RoomPaginationStatus};
 pub use room::{RoomEventCache, RoomEventCacheSubscriber, ThreadEventCacheUpdate};
@@ -744,10 +742,9 @@ impl EventCache {
                     let mut search_index_guard = client.search_index().lock().await;
 
                     for event in timeline_events {
-                        if let Some(index_operation) =
-                            search::parse_timeline_event(&room_cache, &event, &redaction_rules)
-                                .await
-                            && let Err(err) = search_index_guard.execute(index_operation, &room_id)
+                        if let Err(err) = search_index_guard
+                            .handle_timeline_event(&event, &room_cache, &room_id, &redaction_rules)
+                            .await
                         {
                             warn!("Failed to handle event for indexing: {err}")
                         }
