+++ crates/matrix-sdk/src/lib.rs
@@ -67,7 +67,7 @@ pub mod widget;
 pub use account::Account;
 pub use authentication::{AuthApi, AuthSession, SessionTokens};
 #[cfg(feature = "experimental-search")]
-pub use client::search::SearchIndexStoreKind;
+pub mod search_index;
 pub use client::{
     Client, ClientBuildError, ClientBuilder, LoopCtrl, ServerVendorInfo, SessionChange,
     sanitize_server_name,
