+++ crates/matrix-sdk/src/search_index/mod.rs
@@ -12,15 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! The search index is an abstraction layer in the matrix-sdk for the
+//! matrix-sdk-search crate. It provides a [`SearchIndex`] which wraps
+//! multiple [`RoomIndex`].
+
 use std::{collections::hash_map::HashMap, path::PathBuf, sync::Arc};
 
+use matrix_sdk_base::deserialized_responses::TimelineEvent;
 use matrix_sdk_search::{
     error::IndexError,
     index::{RoomIndex, RoomIndexOperation},
 };
-use ruma::{OwnedEventId, OwnedRoomId, RoomId};
+use ruma::{
+    EventId, OwnedEventId, OwnedRoomId, RoomId,
+    events::{
+        AnySyncMessageLikeEvent, AnySyncTimelineEvent,
+        room::{
+            message::{OriginalSyncRoomMessageEvent, Relation, SyncRoomMessageEvent},
+            redaction::SyncRoomRedactionEvent,
+        },
+    },
+    room_version_rules::RedactionRules,
+};
 use tokio::sync::{Mutex, MutexGuard};
-use tracing::{debug, error};
+use tracing::{debug, error, warn};
+
+use crate::event_cache::RoomEventCache;
 
 /// Type of location to store [`RoomIndex`]
 #[derive(Clone, Debug)]
@@ -33,7 +50,7 @@ pub enum SearchIndexStoreKind {
 
 /// Object that handles inteeraction with [`RoomIndex`]'s for search
 #[derive(Clone, Debug)]
-pub(crate) struct SearchIndex {
+pub struct SearchIndex {
     /// HashMap that links each joined room to its RoomIndex
     room_indexes: Arc<Mutex<HashMap<OwnedRoomId, RoomIndex>>>,
 
@@ -42,7 +59,7 @@ pub(crate) struct SearchIndex {
 }
 
 impl SearchIndex {
-    /// Create a new [`SearchIndexHandler`]
+    /// Create a new [`SearchIndex`]
     pub fn new(
         room_indexes: Arc<Mutex<HashMap<OwnedRoomId, RoomIndex>>>,
         search_index_store_kind: SearchIndexStoreKind,
@@ -50,6 +67,7 @@ impl SearchIndex {
         Self { room_indexes, search_index_store_kind }
     }
 
+    /// Aquire [`SearchIndexGuard`] for this [`SearchIndex`].
     pub async fn lock(&self) -> SearchIndexGuard<'_> {
         SearchIndexGuard {
             index_map: self.room_indexes.lock().await,
@@ -58,7 +76,9 @@ impl SearchIndex {
     }
 }
 
-pub(crate) struct SearchIndexGuard<'a> {
+/// Object that represents an aquired [`SearchIndex`].
+#[derive(Debug)]
+pub struct SearchIndexGuard<'a> {
     /// Guard around the [`RoomIndex`] map
     index_map: MutexGuard<'a, HashMap<OwnedRoomId, RoomIndex>>,
 
@@ -136,6 +156,137 @@ impl SearchIndexGuard<'_> {
             let _ = index.commit_and_reload().inspect_err(|err| {
                 error!("error occurred while committing: {err:?}");
             });
+        }
+    }
+
+    /// Given a [`TimelineEvent`] this function will derive a
+    /// [`RoomIndexOperation`], if it should be handled, and execute it;
+    /// returning the result.
+    pub async fn handle_timeline_event(
+        &mut self,
+        event: &TimelineEvent,
+        room_cache: &RoomEventCache,
+        room_id: &RoomId,
+        redaction_rules: &RedactionRules,
+    ) -> Result<(), IndexError> {
+        if let Some(index_operation) =
+            parse_timeline_event(room_cache, event, redaction_rules).await
+        {
+            self.execute(index_operation, room_id)
+        } else {
+            Ok(())
+        }
+    }
+}
+
+/// Given an event id this function returns the most recent edit on said event
+/// or the event itself if there are no edits.
+async fn get_most_recent_edit(
+    cache: &RoomEventCache,
+    original: &EventId,
+) -> Option<OriginalSyncRoomMessageEvent> {
+    use ruma::events::{AnySyncTimelineEvent, relation::RelationType};
+
+    let Some((original_ev, related)) =
+        cache.find_event_with_relations(original, Some(vec![RelationType::Replacement])).await
+    else {
+        debug!("Couldn't find relations for {}", original);
+        return None;
+    };
+
+    match related.last().unwrap_or(&original_ev).raw().deserialize() {
+        Ok(AnySyncTimelineEvent::MessageLike(AnySyncMessageLikeEvent::RoomMessage(latest))) => {
+            latest.as_original().cloned()
+        }
+        _ => None,
+    }
+}
+
+/// If the given [`OriginalSyncRoomMessageEvent`] is an edit we make an
+/// [`RoomIndexOperation::Edit`] with the new most recent version of the
+/// original.
+async fn handle_possible_edit(
+    event: &OriginalSyncRoomMessageEvent,
+    cache: &RoomEventCache,
+) -> Option<RoomIndexOperation> {
+    if let Some(Relation::Replacement(replacement_data)) = &event.content.relates_to {
+        if let Some(recent) = get_most_recent_edit(cache, &replacement_data.event_id).await {
+            return Some(RoomIndexOperation::Edit(replacement_data.event_id.clone(), recent));
+        } else {
+            return Some(RoomIndexOperation::Noop);
+        }
+    }
+    None
+}
+
+/// Return a [`RoomIndexOperation::Edit`] or [`RoomIndexOperation::Add`]
+/// depending on the message.
+async fn handle_room_message(
+    event: SyncRoomMessageEvent,
+    cache: &RoomEventCache,
+) -> Option<RoomIndexOperation> {
+    if let Some(event) = event.as_original() {
+        return handle_possible_edit(event, cache).await.or(get_most_recent_edit(
+            cache,
+            &event.event_id,
+        )
+        .await
+        .map(RoomIndexOperation::Add));
+    }
+    None
+}
+
+/// Return a [`RoomIndexOperation::Edit`] or [`RoomIndexOperation::Remove`]
+/// depending on the message.
+async fn handle_room_redaction(
+    event: SyncRoomRedactionEvent,
+    cache: &RoomEventCache,
+    rules: &RedactionRules,
+) -> Option<RoomIndexOperation> {
+    if let Some(redacted_event_id) = event.redacts(rules)
+        && let Some(redacted_event) = cache.find_event(redacted_event_id).await
+        && let Ok(AnySyncTimelineEvent::MessageLike(AnySyncMessageLikeEvent::RoomMessage(
+            redacted_event,
+        ))) = redacted_event.raw().deserialize()
+        && let Some(redacted_event) = redacted_event.as_original()
+    {
+        return handle_possible_edit(redacted_event, cache)
+            .await
+            .or(Some(RoomIndexOperation::Remove(redacted_event.event_id.clone())));
+    }
+    None
+}
+
+/// Prepare a [`TimelineEvent`] into a [`RoomIndexOperation`] for search
+/// indexing.
+async fn parse_timeline_event(
+    cache: &RoomEventCache,
+    event: &TimelineEvent,
+    redaction_rules: &RedactionRules,
+) -> Option<RoomIndexOperation> {
+    use ruma::events::AnySyncTimelineEvent;
+
+    if event.kind.is_utd() {
+        return None;
+    }
+
+    match event.raw().deserialize() {
+        Ok(event) => match event {
+            AnySyncTimelineEvent::MessageLike(event) => match event {
+                AnySyncMessageLikeEvent::RoomMessage(event) => {
+                    handle_room_message(event, cache).await
+                }
+                AnySyncMessageLikeEvent::RoomRedaction(event) => {
+                    handle_room_redaction(event, cache, redaction_rules).await
+                }
+                _ => None,
+            },
+            AnySyncTimelineEvent::State(_) => None,
+        },
+
+        Err(e) => {
+            warn!("failed to parse event: {e:?}");
+            None
         }
     }
 }
